% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxly.R
\name{boxly}
\alias{boxly}
\title{Create an interactive box plot}
\usage{
boxly(
  outdata,
  color = NULL,
  hover_summary_var = c("n", "min", "q1", "median", "mean", "q3", "max"),
  hover_outlier_label = c("Participant ID", "Parameter value"),
  x_label = "Visit",
  y_label = "Change",
  heading_select_list = "Lab parameter",
  heading_summary_table = "Number of Participants"
)
}
\arguments{
\item{outdata}{An \code{outdata} object created from \code{prepare_ae_forestly()}.}

\item{color}{Color for box plot.}

\item{hover_summary_var}{A character vector of statistics to be displayed
on hover label of box.}

\item{hover_outlier_label}{A character vector of hover label for outlier.
A label from an input data is used if \code{NA} for a variable is specified.}

\item{x_label}{x-axis label.}

\item{y_label}{y-axis label.}

\item{heading_select_list}{Select list menu label.}

\item{heading_summary_table}{Summary table label.}
}
\value{
Interactive box plot.
}
\description{
Create an interactive box plot
}
\examples{
# Only run this example in interactive R sessions
if (interactive()) {
  analysis_plan <- metalite::plan(
    analysis = "boxly",
    population = "apat",
    observation = "wk12",
    parameter = "SODIUM;BILI"
  )

  meta <- metalite::meta_adam(
    population = boxly_adsl,
    observation = boxly_adlb
  ) |>
    metalite::define_plan(analysis_plan) |>
    metalite::define_population(
      name = "apat",
      group = "TRTA",
      subset = SAFFL == "Y",
      label = "Safety Population"
    ) |>
    metalite::define_observation(
      name = "wk12",
      group = "TRTA",
      var = "PARAM",
      subset = AVISITN <= 12 & !is.na(CHG),
      label = "Weeks 0 to 12"
    ) |>
    metalite::define_parameter(
      name = "SODIUM",
      label = "Sodium (mmol/L)",
      subset = PARAMCD == "SODIUM"
    ) |>
    metalite::define_parameter(
      name = "BILI",
      label = "Bilirubin (mg/dL)",
      subset = PARAMCD == "BILI"
    ) |>
    metalite::define_analysis(
      name = "boxly",
      label = "Interactive Box Plot",
      x = "AVISITN",
      y = "CHG"
    ) |>
    metalite::meta_build()

  meta |>
    prepare_boxly() |>
    boxly()
}
}
