% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boxly_adlb}
\alias{boxly_adlb}
\title{A example ADLB dataset}
\format{
A data frame with 24746 and 24 variables:
\describe{
\item{usubjid}{Unique Subject Identifier}
\item{trta}{Actual Treatment}
\item{trtan}{Actual Treatment (N)}
\item{age}{Age}
\item{race}{Race}
\item{sex}{Sex}
\item{saffl}{Safety Population Flag}
\item{avisit}{Analysis Visit}
\item{avisitn}{Analysis Visit (N)}
\item{ady}{Analysis Relative Day}
\item{adt}{Analysis Date}
\item{visit}{Visit Name}
\item{visitnum}{Visit Number}
\item{param}{Parameter}
\item{paramcd}{Parameter Code}
\item{paramn}{Parameter (N)}
\item{aval}{Analysis Value}
\item{base}{Baseline Value}
\item{chg}{Change from Baseline}
\item{anl01fl}{Analysis Record Flag 1}
\item{albtrval}{Amount Threshold Range}
\item{anrind}{Analysis Reference Range Indicator}
\item{bnrind}{Baseline Reference Range Indicator}
\item{ablfl}{Baseline Record Flag}
}
}
\usage{
boxly_adlb
}
\description{
A example ADLB dataset
}
\keyword{datasets}
