% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_boxly.R
\name{meta_boxly}
\alias{meta_boxly}
\title{Create an example metadata object}
\usage{
meta_boxly()
}
\value{
Metadata for creating interactive box plot.
}
\description{
Create an example metadata object
}
\examples{
meta_boxly()
}
