% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boxly_adeg}
\alias{boxly_adeg}
\title{An example ADEG dataset}
\format{
A data frame with 32139 and 35 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
boxly_adeg
}
\description{
Definition of each variable can be found in
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\keyword{datasets}
