% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boxly_adlb}
\alias{boxly_adlb}
\title{An example ADLB dataset}
\format{
A data frame with 24746 and 24 variables:
\describe{
\item{USUBJID}{Unique Subject Identifier}
\item{TRTA}{Actual Treatment}
\item{TRTAN}{Actual Treatment (N)}
\item{AGE}{Age}
\item{RACE}{Race}
\item{SEX}{Sex}
\item{SAFFL}{Safety Population Flag}
\item{AVISIT}{Analysis Visit}
\item{AVISITN}{Analysis Visit (N)}
\item{ADY}{Analysis Relative Day}
\item{ADT}{Analysis Date}
\item{VISIT}{Visit Name}
\item{VISITNUM}{Visit Number}
\item{PARAM}{Parameter}
\item{PARAMCD}{Parameter Code}
\item{PARAMN}{Parameter (N)}
\item{AVAL}{Analysis Value}
\item{BASE}{Baseline Value}
\item{CHG}{Change from Baseline}
\item{ANL01FL}{Analysis Record Flag 1}
\item{ALBTRVAL}{Amount Threshold Range}
\item{ANRIND}{Analysis Reference Range Indicator}
\item{BNRIND}{Baseline Reference Range Indicator}
\item{ABLFL}{Baseline Record Flag}
}
}
\usage{
boxly_adlb
}
\description{
An example ADLB dataset
}
\keyword{datasets}
