% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boxly_adlb}
\alias{boxly_adlb}
\title{An example ADLB dataset}
\format{
A data frame with 24746 and 24 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
boxly_adlb
}
\description{
Definition of each variable can be found in
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\keyword{datasets}
