% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boxly_advs}
\alias{boxly_advs}
\title{An example ADVS dataset}
\format{
A data frame with 32139 and 34 variables:
}
\source{
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\usage{
boxly_advs
}
\description{
Definition of each variable can be found in
\url{https://github.com/phuse-org/phuse-scripts/tree/master/data/adam/cdisc}
}
\keyword{datasets}
