% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_boxly.R
\name{prepare_boxly}
\alias{prepare_boxly}
\title{Prepare data for interactive box plot}
\usage{
prepare_boxly(meta, population, observation, analysis, parameter)
}
\arguments{
\item{meta}{A metadata object created by metalite.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{analysis}{Name of analysis plan.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}
}
\value{
Metadata list with plotting dataset.

Metadata list with plotting dataset
}
\description{
Prepare data for interactive box plot
}
\examples{
library(metalite)

meta <- meta_boxly()
prepare_boxly(meta,
  population = "apat",
  observation = "wk12",
  analysis = "lb_boxly",
  parameter = "sodium;bili;urate"
)
}
